fix: close PR 2697 review gaps
