+++ apps/packages/ui/src/components/Option/Playground/hooks/__tests__/usePlaygroundQueueManagement.document-processing.test.tsx
@@ -177,6 +177,7 @@ describe("usePlaygroundQueueManagement document processing", () => {
         uploadedFiles: [],
         ragMediaIds: [42],
         fileRetrievalEnabled: true,
+        documentSnippetForModel: "parsed notes",
         documentProcessing: turnMetadata
       },
       turnMetadata
@@ -226,6 +227,7 @@ describe("usePlaygroundQueueManagement document processing", () => {
           uploadedFiles: [],
           ragMediaIds: [42],
           fileRetrievalEnabled: true,
+          messageForModel: "summarize\n\nparsed notes",
           userMetadataExtra: {
             documentProcessing: turnMetadata
           }
