+++ tldw_Server_API/app/api/v1/endpoints/media/document_upload_processing.py
@@ -1,10 +1,11 @@
+"""API routes for chat document upload preflight and handoff drafts."""
+
 from __future__ import annotations
 
 import json
 from datetime import datetime, timedelta, timezone
-from pathlib import Path
 from threading import RLock
-from typing import Any, Literal
+from typing import Any
 from uuid import uuid4
 
 from fastapi import APIRouter, Depends, HTTPException, Response, status
@@ -14,13 +15,14 @@ from tldw_Server_API.app.api.v1.schemas.document_upload_processing import (
     ChatDocumentDraftCreateRequest,
     ChatDocumentDraftCreateResponse,
     ChatDocumentDraftReadResponse,
-    DocumentModeCapability,
-    DocumentProcessingMode,
-    DocumentProcessingStatus,
-    DocumentUploadPreflightFile,
-    DocumentUploadPreflightItem,
     DocumentUploadPreflightRequest,
     DocumentUploadPreflightResponse,
+)
+from tldw_Server_API.app.core.Ingestion_Media_Processing.document_upload_preflight import (
+    DEFAULT_MAX_CHAT_UPLOAD_BYTES,
+    DEFAULT_MAX_CHAT_UPLOAD_PAGES,
+    DEFAULT_MAX_DIRECT_CHAT_TOKENS,
+    preflight_document_upload_files,
 )
 from tldw_Server_API.app.core.Ingestion_Media_Processing.OCR.registry import (
     list_backends as _list_backends,
@@ -28,24 +30,6 @@ from tldw_Server_API.app.core.Ingestion_Media_Processing.OCR.registry import (
 
 router = APIRouter(tags=["Media Processing"])
 
-SUPPORTED_DOCUMENT_EXTENSIONS = {
-    ".txt",
-    ".md",
-    ".markdown",
-    ".docx",
-    ".rtf",
-    ".html",
-    ".htm",
-    ".xhtml",
-    ".xml",
-    ".json",
-}
-SUPPORTED_EBOOK_EXTENSIONS = {".epub"}
-SUPPORTED_PDF_EXTENSIONS = {".pdf"}
-
-DEFAULT_MAX_CHAT_UPLOAD_BYTES = 20 * 1024 * 1024
-DEFAULT_MAX_CHAT_UPLOAD_PAGES = 200
-DEFAULT_MAX_DIRECT_CHAT_TOKENS = 24_000
 DRAFT_TTL_SECONDS = 15 * 60
 MAX_DRAFT_PAYLOAD_BYTES = DEFAULT_MAX_CHAT_UPLOAD_BYTES
 MAX_DRAFTS_TOTAL = 256
@@ -97,149 +81,6 @@ def _enforce_draft_quota(owner: str) -> None:
         )
 
 
-MediaType = Literal["pdf", "document", "ebook", "unsupported"]
-
-
-def _media_type_for(filename: str) -> MediaType:
-    suffix = Path(filename).suffix.lower()
-    if suffix in SUPPORTED_PDF_EXTENSIONS:
-        return "pdf"
-    if suffix in SUPPORTED_DOCUMENT_EXTENSIONS:
-        return "document"
-    if suffix in SUPPORTED_EBOOK_EXTENSIONS:
-        return "ebook"
-    return "unsupported"
-
-
-def _capability(
-    available: bool,
-    status_value: DocumentProcessingStatus,
-    reason: str | None = None,
-) -> DocumentModeCapability:
-    return DocumentModeCapability(
-        available=available,
-        status=status_value,
-        reason=reason,
-    )
-
-
-def _available(reason: str | None = None) -> DocumentModeCapability:
-    return _capability(True, "available", reason)
-
-
-def _unavailable(reason: str) -> DocumentModeCapability:
-    return _capability(False, "unavailable", reason)
-
-
-def _blocked(reason: str) -> DocumentModeCapability:
-    return _capability(False, "blocked", reason)
-
-
-def _ocr_available() -> bool:
-    backends = _list_ocr_backends()
-    return any(
-        isinstance(details, dict) and details.get("available") is True
-        for details in backends.values()
-    )
-
-
-def _size_limit_reason(file: DocumentUploadPreflightFile) -> str | None:
-    if file.size_bytes <= DEFAULT_MAX_CHAT_UPLOAD_BYTES:
-        return None
-    size_mb = DEFAULT_MAX_CHAT_UPLOAD_BYTES // (1024 * 1024)
-    return f"{file.filename} exceeds {size_mb} MB limit"
-
-
-def _page_limit_reason(file: DocumentUploadPreflightFile) -> str | None:
-    if file.page_count is None or file.page_count <= DEFAULT_MAX_CHAT_UPLOAD_PAGES:
-        return None
-    return f"{file.filename} exceeds {DEFAULT_MAX_CHAT_UPLOAD_PAGES} page limit"
-
-
-def _token_limit_reason(file: DocumentUploadPreflightFile) -> str | None:
-    if file.estimated_tokens is None or file.estimated_tokens <= DEFAULT_MAX_DIRECT_CHAT_TOKENS:
-        return None
-    return f"{file.filename} exceeds {DEFAULT_MAX_DIRECT_CHAT_TOKENS} token direct-chat limit"
-
-
-def _unsupported_modes() -> dict[DocumentProcessingMode, DocumentModeCapability]:
-    reason = "Unsupported document type"
-    return {
-        "add_to_chat": _unavailable(reason),
-        "ocr_pages": _unavailable(reason),
-        "ingest_to_library": _unavailable(reason),
-    }
-
-
-def _preflight_file(file: DocumentUploadPreflightFile) -> DocumentUploadPreflightItem:
-    media_type = _media_type_for(file.filename)
-    if media_type == "unsupported":
-        return DocumentUploadPreflightItem(
-            client_id=file.client_id,
-            filename=file.filename,
-            media_type="unsupported",
-            default_mode=None,
-            modes=_unsupported_modes(),
-            max_size_bytes=DEFAULT_MAX_CHAT_UPLOAD_BYTES,
-            max_pages=DEFAULT_MAX_CHAT_UPLOAD_PAGES,
-            max_chat_tokens=DEFAULT_MAX_DIRECT_CHAT_TOKENS,
-            estimated_pages=file.page_count,
-            estimated_tokens=file.estimated_tokens,
-            requires_send_time_estimate=False,
-        )
-
-    size_reason = _size_limit_reason(file)
-    page_reason = _page_limit_reason(file)
-    token_reason = _token_limit_reason(file)
-    direct_reason = size_reason or page_reason or token_reason
-
-    add_to_chat = _blocked(direct_reason) if direct_reason else _available()
-    ingest = _blocked(size_reason) if size_reason else _available()
-    ocr = _ocr_capability(file, media_type, size_reason or page_reason)
-
-    modes: dict[DocumentProcessingMode, DocumentModeCapability] = {
-        "add_to_chat": add_to_chat,
-        "ocr_pages": ocr,
-        "ingest_to_library": ingest,
-    }
-    default_mode = next(
-        (
-            mode
-            for mode in ("add_to_chat", "ingest_to_library", "ocr_pages")
-            if modes[mode].available
-        ),
-        None,
-    )
-    return DocumentUploadPreflightItem(
-        client_id=file.client_id,
-        filename=file.filename,
-        media_type=media_type,
-        default_mode=default_mode,
-        modes=modes,
-        max_size_bytes=DEFAULT_MAX_CHAT_UPLOAD_BYTES,
-        max_pages=DEFAULT_MAX_CHAT_UPLOAD_PAGES,
-        max_chat_tokens=DEFAULT_MAX_DIRECT_CHAT_TOKENS,
-        estimated_pages=file.page_count,
-        estimated_tokens=file.estimated_tokens,
-        requires_send_time_estimate=file.page_count is None or file.estimated_tokens is None,
-    )
-
-
-def _ocr_capability(
-    file: DocumentUploadPreflightFile,
-    media_type: str,
-    blocked_reason: str | None,
-) -> DocumentModeCapability:
-    if media_type != "pdf":
-        suffix = Path(file.filename).suffix.upper() or "file"
-        return _unavailable(f"OCR unavailable: server cannot render {suffix} pages")
-    if blocked_reason:
-        return _blocked(blocked_reason)
-    if not _ocr_available():
-        return _unavailable("OCR unavailable: no OCR backend configured")
-    return _available()
-
-
 def _draft_for(draft_id: str, current_user: Any) -> dict[str, Any]:
     with _DRAFTS_LOCK:
         _cleanup_expired_drafts()
@@ -259,13 +100,17 @@ def preflight_document_upload(
     _current_user: Any = Depends(get_request_user),
 ) -> DocumentUploadPreflightResponse:
     return DocumentUploadPreflightResponse(
-        files=[_preflight_file(file) for file in request.files],
+        files=preflight_document_upload_files(
+            request.files,
+            list_ocr_backends=_list_ocr_backends,
+        ),
     )
 
 
 @router.post(
     "/document-upload/drafts",
     response_model=ChatDocumentDraftCreateResponse,
+    dependencies=[Depends(rbac_rate_limit("media.write"))],
 )
 def create_document_upload_draft(
     request: ChatDocumentDraftCreateRequest,
@@ -300,6 +145,7 @@ def create_document_upload_draft(
 @router.get(
     "/document-upload/drafts/{draft_id}",
     response_model=ChatDocumentDraftReadResponse,
+    dependencies=[Depends(rbac_rate_limit("media.read"))],
 )
 def read_document_upload_draft(
     draft_id: str,
@@ -317,6 +163,7 @@ def read_document_upload_draft(
 @router.delete(
     "/document-upload/drafts/{draft_id}",
     status_code=status.HTTP_204_NO_CONTENT,
+    dependencies=[Depends(rbac_rate_limit("media.write"))],
 )
 def delete_document_upload_draft(
     draft_id: str,
